Version 3.0
I cannot thank Suraj N. Kurapati enough for submitting a patch that sets the                                                                                                        '[ and '] marks around the FIGlet-ified text. Thank you for your great work. If it weren't for you there wouldn't be a version 3.0!

This version also changes the spelling of the commands from :Figlet* to :FIGlet* in keeping with the style of www.figlet.org.